@@ -20,23 +20,23 @@
   },
   "homepage": "https://github.com/im6/snake-reactive#readme",
   "dependencies": {
-    "css-loader": "^6.6.0",
-    "less": "^4.1.2",
-    "less-loader": "^10.2.0",
-    "rxjs": "^7.5.4",
+    "css-loader": "^6.7.2",
+    "less": "^4.1.3",
+    "less-loader": "^11.1.0",
+    "rxjs": "^7.6.0",
     "style-loader": "^3.3.1",
-    "ts-loader": "^9.2.6",
-    "typescript": "^4.5.5",
-    "webpack": "^5.69.1",
-    "webpack-cli": "^4.9.2",
-    "webpack-dev-server": "^4.7.4"
+    "ts-loader": "^9.4.2",
+    "typescript": "^4.9.3",
+    "webpack": "^5.75.0",
+    "webpack-cli": "^5.0.0",
+    "webpack-dev-server": "^4.11.1"
   },
   "devDependencies": {
-    "@types/jest": "^27.4.1",
-    "@typescript-eslint/eslint-plugin": "^5.12.1",
-    "@typescript-eslint/parser": "^5.12.1",
-    "eslint": "^8.9.0",
-    "jest": "^27.5.1",
-    "ts-jest": "^27.1.3"
+    "@types/jest": "^29.2.3",
+    "@typescript-eslint/eslint-plugin": "^5.45.0",
+    "@typescript-eslint/parser": "^5.45.0",
+    "eslint": "^8.29.0",
+    "jest": "^29.3.1",
+    "ts-jest": "^29.0.3"
   }
 }
