@@ -1,2 +1,1 @@
-/*! For license information please see bundle.js.LICENSE.txt */
-(()=>{"use strict";var t={935:(t,n,r)=>{r.d(n,{Z:()=>c});var e=r(81),o=r.n(e),i=r(645),u=r.n(i)()(o());u.push([t.id,'h1,\nh2,\nh3,\nh4,\nh5,\nspan,\na,\np,\nbutton {\n  font-family: "Play", sans-serif;\n}\nheader h1 {\n  color: #ff7148;\n  text-align: center;\n}\ncanvas {\n  display: block;\n  border: 5px solid #93b1c6;\n  margin: 20px auto;\n  width: 400px;\n  height: 400px;\n}\n.gameControl {\n  text-align: center;\n  display: none;\n}\n#gameStartBtn {\n  padding: 8px 50px;\n  font-size: 1.2em;\n}\n#gameStartBtn:focus {\n  outline: none;\n}\n',""]);const c=u},645:t=>{t.exports=function(t){var n=[];return n.toString=function(){return this.map((function(n){var r="",e=void 0!==n[5];return n[4]&&(r+="@supports (".concat(n[4],") {")),n[2]&&(r+="@media ".concat(n[2]," {")),e&&(r+="@layer".concat(n[5].length>0?" ".concat(n[5]):""," {")),r+=t(n),e&&(r+="}"),n[2]&&(r+="}"),n[4]&&(r+="}"),r})).join("")},n.i=function(t,r,e,o,i){"string"==typeof t&&(t=[[null,t,void 0]]);var u={};if(e)for(var c=0;c<this.length;c++){var a=this[c][0];null!=a&&(u[a]=!0)}for(var s=0;s<t.length;s++){var l=[].concat(t[s]);e&&u[l[0]]||(void 0!==i&&(void 0===l[5]||(l[1]="@layer".concat(l[5].length>0?" ".concat(l[5]):""," {").concat(l[1],"}")),l[5]=i),r&&(l[2]?(l[1]="@media ".concat(l[2]," {").concat(l[1],"}"),l[2]=r):l[2]=r),o&&(l[4]?(l[1]="@supports (".concat(l[4],") {").concat(l[1],"}"),l[4]=o):l[4]="".concat(o)),n.push(l))}},n}},81:t=>{t.exports=function(t){return t[1]}},379:t=>{var n=[];function r(t){for(var r=-1,e=0;e<n.length;e++)if(n[e].identifier===t){r=e;break}return r}function e(t,e){for(var i={},u=[],c=0;c<t.length;c++){var a=t[c],s=e.base?a[0]+e.base:a[0],l=i[s]||0,f="".concat(s," ").concat(l);i[s]=l+1;var p=r(f),h={css:a[1],media:a[2],sourceMap:a[3],supports:a[4],layer:a[5]};if(-1!==p)n[p].references++,n[p].updater(h);else{var d=o(h,e);e.byIndex=c,n.splice(c,0,{identifier:f,updater:d,references:1})}u.push(f)}return u}function o(t,n){var r=n.domAPI(n);return r.update(t),function(n){if(n){if(n.css===t.css&&n.media===t.media&&n.sourceMap===t.sourceMap&&n.supports===t.supports&&n.layer===t.layer)return;r.update(t=n)}else r.remove()}}t.exports=function(t,o){var i=e(t=t||[],o=o||{});return function(t){t=t||[];for(var u=0;u<i.length;u++){var c=r(i[u]);n[c].references--}for(var a=e(t,o),s=0;s<i.length;s++){var l=r(i[s]);0===n[l].references&&(n[l].updater(),n.splice(l,1))}i=a}}},569:t=>{var n={};t.exports=function(t,r){var e=function(t){if(void 0===n[t]){var r=document.querySelector(t);if(window.HTMLIFrameElement&&r instanceof window.HTMLIFrameElement)try{r=r.contentDocument.head}catch(t){r=null}n[t]=r}return n[t]}(t);if(!e)throw new Error("Couldn't find a style target. This probably means that the value for the 'insert' parameter is invalid.");e.appendChild(r)}},216:t=>{t.exports=function(t){var n=document.createElement("style");return t.setAttributes(n,t.attributes),t.insert(n,t.options),n}},565:(t,n,r)=>{t.exports=function(t){var n=r.nc;n&&t.setAttribute("nonce",n)}},795:t=>{t.exports=function(t){var n=t.insertStyleElement(t);return{update:function(r){!function(t,n,r){var e="";r.supports&&(e+="@supports (".concat(r.supports,") {")),r.media&&(e+="@media ".concat(r.media," {"));var o=void 0!==r.layer;o&&(e+="@layer".concat(r.layer.length>0?" ".concat(r.layer):""," {")),e+=r.css,o&&(e+="}"),r.media&&(e+="}"),r.supports&&(e+="}");var i=r.sourceMap;i&&"undefined"!=typeof btoa&&(e+="\n/*# sourceMappingURL=data:application/json;base64,".concat(btoa(unescape(encodeURIComponent(JSON.stringify(i))))," */")),n.styleTagTransform(e,t,n.options)}(n,t,r)},remove:function(){!function(t){if(null===t.parentNode)return!1;t.parentNode.removeChild(t)}(n)}}}},589:t=>{t.exports=function(t,n){if(n.styleSheet)n.styleSheet.cssText=t;else{for(;n.firstChild;)n.removeChild(n.firstChild);n.appendChild(document.createTextNode(t))}}}},n={};function r(e){var o=n[e];if(void 0!==o)return o.exports;var i=n[e]={id:e,exports:{}};return t[e](i,i.exports,r),i.exports}r.n=t=>{var n=t&&t.__esModule?()=>t.default:()=>t;return r.d(n,{a:n}),n},r.d=(t,n)=>{for(var e in n)r.o(n,e)&&!r.o(t,e)&&Object.defineProperty(t,e,{enumerable:!0,get:n[e]})},r.o=(t,n)=>Object.prototype.hasOwnProperty.call(t,n),(()=>{var t=r(379),n=r.n(t),e=r(795),o=r.n(e),i=r(569),u=r.n(i),c=r(565),a=r.n(c),s=r(216),l=r.n(s),f=r(589),p=r.n(f),h=r(935),d={};d.styleTagTransform=p(),d.setAttributes=a(),d.insert=u().bind(null,"head"),d.domAPI=o(),d.insertStyleElement=l(),n()(h.Z,d),h.Z&&h.Z.locals&&h.Z.locals;var v=function(t,n){return v=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(t,n){t.__proto__=n}||function(t,n){for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(t[r]=n[r])},v(t,n)};function y(t,n){if("function"!=typeof n&&null!==n)throw new TypeError("Class extends value "+String(n)+" is not a constructor or null");function r(){this.constructor=t}v(t,n),t.prototype=null===n?Object.create(n):(r.prototype=n.prototype,new r)}function b(t,n){var r,e,o,i,u={label:0,sent:function(){if(1&o[0])throw o[1];return o[1]},trys:[],ops:[]};return i={next:c(0),throw:c(1),return:c(2)},"function"==typeof Symbol&&(i[Symbol.iterator]=function(){return this}),i;function c(i){return function(c){return function(i){if(r)throw new TypeError("Generator is already executing.");for(;u;)try{if(r=1,e&&(o=2&i[0]?e.return:i[0]?e.throw||((o=e.return)&&o.call(e),0):e.next)&&!(o=o.call(e,i[1])).done)return o;switch(e=0,o&&(i=[2&i[0],o.value]),i[0]){case 0:case 1:o=i;break;case 4:return u.label++,{value:i[1],done:!1};case 5:u.label++,e=i[1],i=[0];continue;case 7:i=u.ops.pop(),u.trys.pop();continue;default:if(!((o=(o=u.trys).length>0&&o[o.length-1])||6!==i[0]&&2!==i[0])){u=0;continue}if(3===i[0]&&(!o||i[1]>o[0]&&i[1]<o[3])){u.label=i[1];break}if(6===i[0]&&u.label<o[1]){u.label=o[1],o=i;break}if(o&&u.label<o[2]){u.label=o[2],u.ops.push(i);break}o[2]&&u.ops.pop(),u.trys.pop();continue}i=n.call(t,u)}catch(t){i=[6,t],e=0}finally{r=o=0}if(5&i[0])throw i[1];return{value:i[0]?i[1]:void 0,done:!0}}([i,c])}}}function m(t){var n="function"==typeof Symbol&&Symbol.iterator,r=n&&t[n],e=0;if(r)return r.call(t);if(t&&"number"==typeof t.length)return{next:function(){return t&&e>=t.length&&(t=void 0),{value:t&&t[e++],done:!t}}};throw new TypeError(n?"Object is not iterable.":"Symbol.iterator is not defined.")}function w(t,n){var r="function"==typeof Symbol&&t[Symbol.iterator];if(!r)return t;var e,o,i=r.call(t),u=[];try{for(;(void 0===n||n-- >0)&&!(e=i.next()).done;)u.push(e.value)}catch(t){o={error:t}}finally{try{e&&!e.done&&(r=i.return)&&r.call(i)}finally{if(o)throw o.error}}return u}function x(t,n,r){if(r||2===arguments.length)for(var e,o=0,i=n.length;o<i;o++)!e&&o in n||(e||(e=Array.prototype.slice.call(n,0,o)),e[o]=n[o]);return t.concat(e||Array.prototype.slice.call(n))}function g(t){return this instanceof g?(this.v=t,this):new g(t)}function _(t,n,r){if(!Symbol.asyncIterator)throw new TypeError("Symbol.asyncIterator is not defined.");var e,o=r.apply(t,n||[]),i=[];return e={},u("next"),u("throw"),u("return"),e[Symbol.asyncIterator]=function(){return this},e;function u(t){o[t]&&(e[t]=function(n){return new Promise((function(r,e){i.push([t,n,r,e])>1||c(t,n)}))})}function c(t,n){try{(r=o[t](n)).value instanceof g?Promise.resolve(r.value.v).then(a,s):l(i[0][2],r)}catch(t){l(i[0][3],t)}var r}function a(t){c("next",t)}function s(t){c("throw",t)}function l(t,n){t(n),i.shift(),i.length&&c(i[0][0],i[0][1])}}function S(t){return"function"==typeof t}function A(t){var n=t((function(t){Error.call(t),t.stack=(new Error).stack}));return n.prototype=Object.create(Error.prototype),n.prototype.constructor=n,n}Object.create,Object.create;var E=A((function(t){return function(n){t(this),this.message=n?n.length+" errors occurred during unsubscription:\n"+n.map((function(t,n){return n+1+") "+t.toString()})).join("\n  "):"",this.name="UnsubscriptionError",this.errors=n}}));function I(t,n){if(t){var r=t.indexOf(n);0<=r&&t.splice(r,1)}}var T=function(){function t(t){this.initialTeardown=t,this.closed=!1,this._parentage=null,this._teardowns=null}return t.prototype.unsubscribe=function(){var t,n,r,e,o;if(!this.closed){this.closed=!0;var i=this._parentage;if(i)if(this._parentage=null,Array.isArray(i))try{for(var u=m(i),c=u.next();!c.done;c=u.next())c.value.remove(this)}catch(n){t={error:n}}finally{try{c&&!c.done&&(n=u.return)&&n.call(u)}finally{if(t)throw t.error}}else i.remove(this);var a=this.initialTeardown;if(S(a))try{a()}catch(t){o=t instanceof E?t.errors:[t]}var s=this._teardowns;if(s){this._teardowns=null;try{for(var l=m(s),f=l.next();!f.done;f=l.next()){var p=f.value;try{C(p)}catch(t){o=null!=o?o:[],t instanceof E?o=x(x([],w(o)),w(t.errors)):o.push(t)}}}catch(t){r={error:t}}finally{try{f&&!f.done&&(e=l.return)&&e.call(l)}finally{if(r)throw r.error}}}if(o)throw new E(o)}},t.prototype.add=function(n){var r;if(n&&n!==this)if(this.closed)C(n);else{if(n instanceof t){if(n.closed||n._hasParent(this))return;n._addParent(this)}(this._teardowns=null!==(r=this._teardowns)&&void 0!==r?r:[]).push(n)}},t.prototype._hasParent=function(t){var n=this._parentage;return n===t||Array.isArray(n)&&n.includes(t)},t.prototype._addParent=function(t){var n=this._parentage;this._parentage=Array.isArray(n)?(n.push(t),n):n?[n,t]:t},t.prototype._removeParent=function(t){var n=this._parentage;n===t?this._parentage=null:Array.isArray(n)&&I(n,t)},t.prototype.remove=function(n){var r=this._teardowns;r&&I(r,n),n instanceof t&&n._removeParent(this)},t.EMPTY=((n=new t).closed=!0,n),t;var n}(),O=T.EMPTY;function k(t){return t instanceof T||t&&"closed"in t&&S(t.remove)&&S(t.add)&&S(t.unsubscribe)}function C(t){S(t)?t():t.unsubscribe()}var P=function(t){function n(n,r){return t.call(this)||this}return y(n,t),n.prototype.schedule=function(t,n){return void 0===n&&(n=0),this},n}(T),j={setInterval:function(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];var r=j.delegate;return((null==r?void 0:r.setInterval)||setInterval).apply(void 0,x([],w(t)))},clearInterval:function(t){var n=j.delegate;return((null==n?void 0:n.clearInterval)||clearInterval)(t)},delegate:void 0},F=function(t){function n(n,r){var e=t.call(this,n,r)||this;return e.scheduler=n,e.work=r,e.pending=!1,e}return y(n,t),n.prototype.schedule=function(t,n){if(void 0===n&&(n=0),this.closed)return this;this.state=t;var r=this.id,e=this.scheduler;return null!=r&&(this.id=this.recycleAsyncId(e,r,n)),this.pending=!0,this.delay=n,this.id=this.id||this.requestAsyncId(e,this.id,n),this},n.prototype.requestAsyncId=function(t,n,r){return void 0===r&&(r=0),j.setInterval(t.flush.bind(t,this),r)},n.prototype.recycleAsyncId=function(t,n,r){if(void 0===r&&(r=0),null!=r&&this.delay===r&&!1===this.pending)return n;j.clearInterval(n)},n.prototype.execute=function(t,n){if(this.closed)return new Error("executing a cancelled action");this.pending=!1;var r=this._execute(t,n);if(r)return r;!1===this.pending&&null!=this.id&&(this.id=this.recycleAsyncId(this.scheduler,this.id,null))},n.prototype._execute=function(t,n){var r,e=!1;try{this.work(t)}catch(t){e=!0,r=t||new Error("Scheduled action threw falsy error")}if(e)return this.unsubscribe(),r},n.prototype.unsubscribe=function(){if(!this.closed){var n=this.id,r=this.scheduler,e=r.actions;this.work=this.state=this.scheduler=null,this.pending=!1,I(e,this),null!=n&&(this.id=this.recycleAsyncId(r,n,null)),this.delay=null,t.prototype.unsubscribe.call(this)}},n}(P),B={now:function(){return(B.delegate||Date).now()},delegate:void 0},L=function(){function t(n,r){void 0===r&&(r=t.now),this.schedulerActionCtor=n,this.now=r}return t.prototype.schedule=function(t,n,r){return void 0===n&&(n=0),new this.schedulerActionCtor(this,t).schedule(r,n)},t.now=B.now,t}(),M=function(t){function n(n,r){void 0===r&&(r=L.now);var e=t.call(this,n,r)||this;return e.actions=[],e._active=!1,e._scheduled=void 0,e}return y(n,t),n.prototype.flush=function(t){var n=this.actions;if(this._active)n.push(t);else{var r;this._active=!0;do{if(r=t.execute(t.state,t.delay))break}while(t=n.shift());if(this._active=!1,r){for(;t=n.shift();)t.unsubscribe();throw r}}},n}(L),q=new M(F),z=q,R=null,U=null,N=void 0,Z=!1,G=!1,Y={setTimeout:function(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];var r=Y.delegate;return((null==r?void 0:r.setTimeout)||setTimeout).apply(void 0,x([],w(t)))},clearTimeout:function(t){var n=Y.delegate;return((null==n?void 0:n.clearTimeout)||clearTimeout)(t)},delegate:void 0};function D(t){Y.setTimeout((function(){if(!R)throw t;R(t)}))}function H(){}var V=J("C",void 0,void 0);function J(t,n,r){return{kind:t,value:n,error:r}}var K=null;function Q(t){if(Z){var n=!K;if(n&&(K={errorThrown:!1,error:null}),t(),n){var r=K,e=r.errorThrown,o=r.error;if(K=null,e)throw o}}else t()}var W=function(t){function n(n){var r=t.call(this)||this;return r.isStopped=!1,n?(r.destination=n,k(n)&&n.add(r)):r.destination=ot,r}return y(n,t),n.create=function(t,n,r){return new nt(t,n,r)},n.prototype.next=function(t){this.isStopped?et(function(t){return J("N",t,void 0)}(t),this):this._next(t)},n.prototype.error=function(t){this.isStopped?et(J("E",void 0,t),this):(this.isStopped=!0,this._error(t))},n.prototype.complete=function(){this.isStopped?et(V,this):(this.isStopped=!0,this._complete())},n.prototype.unsubscribe=function(){this.closed||(this.isStopped=!0,t.prototype.unsubscribe.call(this),this.destination=null)},n.prototype._next=function(t){this.destination.next(t)},n.prototype._error=function(t){try{this.destination.error(t)}finally{this.unsubscribe()}},n.prototype._complete=function(){try{this.destination.complete()}finally{this.unsubscribe()}},n}(T),X=Function.prototype.bind;function $(t,n){return X.call(t,n)}var tt=function(){function t(t){this.partialObserver=t}return t.prototype.next=function(t){var n=this.partialObserver;if(n.next)try{n.next(t)}catch(t){rt(t)}},t.prototype.error=function(t){var n=this.partialObserver;if(n.error)try{n.error(t)}catch(t){rt(t)}else rt(t)},t.prototype.complete=function(){var t=this.partialObserver;if(t.complete)try{t.complete()}catch(t){rt(t)}},t}(),nt=function(t){function n(n,r,e){var o,i,u=t.call(this)||this;return S(n)||!n?o={next:null!=n?n:void 0,error:null!=r?r:void 0,complete:null!=e?e:void 0}:u&&G?((i=Object.create(n)).unsubscribe=function(){return u.unsubscribe()},o={next:n.next&&$(n.next,i),error:n.error&&$(n.error,i),complete:n.complete&&$(n.complete,i)}):o=n,u.destination=new tt(o),u}return y(n,t),n}(W);function rt(t){var n;Z?(n=t,Z&&K&&(K.errorThrown=!0,K.error=n)):D(t)}function et(t,n){var r=U;r&&Y.setTimeout((function(){return r(t,n)}))}var ot={closed:!0,next:H,error:function(t){throw t},complete:H},it="function"==typeof Symbol&&Symbol.observable||"@@observable";function ut(t){return t}function ct(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];return at(t)}function at(t){return 0===t.length?ut:1===t.length?t[0]:function(n){return t.reduce((function(t,n){return n(t)}),n)}}var st=function(){function t(t){t&&(this._subscribe=t)}return t.prototype.lift=function(n){var r=new t;return r.source=this,r.operator=n,r},t.prototype.subscribe=function(t,n,r){var e,o=this,i=(e=t)&&e instanceof W||function(t){return t&&S(t.next)&&S(t.error)&&S(t.complete)}(e)&&k(e)?t:new nt(t,n,r);return Q((function(){var t=o,n=t.operator,r=t.source;i.add(n?n.call(i,r):r?o._subscribe(i):o._trySubscribe(i))})),i},t.prototype._trySubscribe=function(t){try{return this._subscribe(t)}catch(n){t.error(n)}},t.prototype.forEach=function(t,n){var r=this;return new(n=lt(n))((function(n,e){var o=new nt({next:function(n){try{t(n)}catch(t){e(t),o.unsubscribe()}},error:e,complete:n});r.subscribe(o)}))},t.prototype._subscribe=function(t){var n;return null===(n=this.source)||void 0===n?void 0:n.subscribe(t)},t.prototype[it]=function(){return this},t.prototype.pipe=function(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];return at(t)(this)},t.prototype.toPromise=function(t){var n=this;return new(t=lt(t))((function(t,r){var e;n.subscribe((function(t){return e=t}),(function(t){return r(t)}),(function(){return t(e)}))}))},t.create=function(n){return new t(n)},t}();function lt(t){var n;return null!==(n=null!=t?t:N)&&void 0!==n?n:Promise}function ft(t){return t&&S(t.schedule)}function pt(t,n){return void 0===t&&(t=0),void 0===n&&(n=q),t<0&&(t=0),function(t,n,r){void 0===t&&(t=0),void 0===r&&(r=z);var e=-1;return null!=n&&(ft(n)?r=n:e=n),new st((function(n){var o,i=(o=t)instanceof Date&&!isNaN(o)?+t-r.now():t;i<0&&(i=0);var u=0;return r.schedule((function(){n.closed||(n.next(u++),0<=e?this.schedule(void 0,e):n.complete())}),i)}))}(t,t,n)}var ht=function(t){return t&&"number"==typeof t.length&&"function"!=typeof t};function dt(t){return S(null==t?void 0:t.then)}function vt(t){return S(t[it])}function yt(t){return Symbol.asyncIterator&&S(null==t?void 0:t[Symbol.asyncIterator])}function bt(t){return new TypeError("You provided "+(null!==t&&"object"==typeof t?"an invalid object":"'"+t+"'")+" where a stream was expected. You can provide an Observable, Promise, ReadableStream, Array, AsyncIterable, or Iterable.")}var mt="function"==typeof Symbol&&Symbol.iterator?Symbol.iterator:"@@iterator";function wt(t){return S(null==t?void 0:t[mt])}function xt(t){return _(this,arguments,(function(){var n,r,e;return b(this,(function(o){switch(o.label){case 0:n=t.getReader(),o.label=1;case 1:o.trys.push([1,,9,10]),o.label=2;case 2:return[4,g(n.read())];case 3:return r=o.sent(),e=r.value,r.done?[4,g(void 0)]:[3,5];case 4:return[2,o.sent()];case 5:return[4,g(e)];case 6:return[4,o.sent()];case 7:return o.sent(),[3,2];case 8:return[3,10];case 9:return n.releaseLock(),[7];case 10:return[2]}}))}))}function gt(t){return S(null==t?void 0:t.getReader)}function _t(t){if(t instanceof st)return t;if(null!=t){if(vt(t))return o=t,new st((function(t){var n=o[it]();if(S(n.subscribe))return n.subscribe(t);throw new TypeError("Provided object does not correctly implement Symbol.observable")}));if(ht(t))return e=t,new st((function(t){for(var n=0;n<e.length&&!t.closed;n++)t.next(e[n]);t.complete()}));if(dt(t))return r=t,new st((function(t){r.then((function(n){t.closed||(t.next(n),t.complete())}),(function(n){return t.error(n)})).then(null,D)}));if(yt(t))return St(t);if(wt(t))return n=t,new st((function(t){var r,e;try{for(var o=m(n),i=o.next();!i.done;i=o.next()){var u=i.value;if(t.next(u),t.closed)return}}catch(t){r={error:t}}finally{try{i&&!i.done&&(e=o.return)&&e.call(o)}finally{if(r)throw r.error}}t.complete()}));if(gt(t))return St(xt(t))}var n,r,e,o;throw bt(t)}function St(t){return new st((function(n){(function(t,n){var r,e,o,i,u,c,a,s;return u=this,c=void 0,s=function(){var u,c;return b(this,(function(a){switch(a.label){case 0:a.trys.push([0,5,6,11]),r=function(t){if(!Symbol.asyncIterator)throw new TypeError("Symbol.asyncIterator is not defined.");var n,r=t[Symbol.asyncIterator];return r?r.call(t):(t=m(t),n={},e("next"),e("throw"),e("return"),n[Symbol.asyncIterator]=function(){return this},n);function e(r){n[r]=t[r]&&function(n){return new Promise((function(e,o){!function(t,n,r,e){Promise.resolve(e).then((function(n){t({value:n,done:r})}),n)}(e,o,(n=t[r](n)).done,n.value)}))}}}(t),a.label=1;case 1:return[4,r.next()];case 2:if((e=a.sent()).done)return[3,4];if(u=e.value,n.next(u),n.closed)return[2];a.label=3;case 3:return[3,1];case 4:return[3,11];case 5:return c=a.sent(),o={error:c},[3,11];case 6:return a.trys.push([6,,9,10]),e&&!e.done&&(i=r.return)?[4,i.call(r)]:[3,8];case 7:a.sent(),a.label=8;case 8:return[3,10];case 9:if(o)throw o.error;return[7];case 10:return[7];case 11:return n.complete(),[2]}}))},new((a=void 0)||(a=Promise))((function(t,n){function r(t){try{o(s.next(t))}catch(t){n(t)}}function e(t){try{o(s.throw(t))}catch(t){n(t)}}function o(n){var o;n.done?t(n.value):(o=n.value,o instanceof a?o:new a((function(t){t(o)}))).then(r,e)}o((s=s.apply(u,c||[])).next())}))})(t,n).catch((function(t){return n.error(t)}))}))}function At(t){return function(n){if(function(t){return S(null==t?void 0:t.lift)}(n))return n.lift((function(n){try{return t(n,this)}catch(t){this.error(t)}}));throw new TypeError("Unable to lift unknown Observable type")}}function Et(t,n,r,e,o){return new It(t,n,r,e,o)}var It=function(t){function n(n,r,e,o,i,u){var c=t.call(this,n)||this;return c.onFinalize=i,c.shouldUnsubscribe=u,c._next=r?function(t){try{r(t)}catch(t){n.error(t)}}:t.prototype._next,c._error=o?function(t){try{o(t)}catch(t){n.error(t)}finally{this.unsubscribe()}}:t.prototype._error,c._complete=e?function(){try{e()}catch(t){n.error(t)}finally{this.unsubscribe()}}:t.prototype._complete,c}return y(n,t),n.prototype.unsubscribe=function(){var n;if(!this.shouldUnsubscribe||this.shouldUnsubscribe()){var r=this.closed;t.prototype.unsubscribe.call(this),!r&&(null===(n=this.onFinalize)||void 0===n||n.call(this))}},n}(W);function Tt(t,n){return At((function(r,e){var o=0;r.subscribe(Et(e,(function(r){e.next(t.call(n,r,o++))})))}))}function Ot(t,n,r,e,o){void 0===e&&(e=0),void 0===o&&(o=!1);var i=n.schedule((function(){r(),o?t.add(this.schedule(null,e)):this.unsubscribe()}),e);if(t.add(i),!o)return i}function kt(t,n,r){return void 0===r&&(r=1/0),S(n)?kt((function(r,e){return Tt((function(t,o){return n(r,t,e,o)}))(_t(t(r,e)))}),r):("number"==typeof n&&(r=n),At((function(n,e){return function(t,n,r,e,o,i,u,c){var a=[],s=0,l=0,f=!1,p=function(){!f||a.length||s||n.complete()},h=function(t){return s<e?d(t):a.push(t)},d=function(t){s++;var o=!1;_t(r(t,l++)).subscribe(Et(n,(function(t){n.next(t)}),(function(){o=!0}),void 0,(function(){if(o)try{s--;for(;a.length&&s<e;)t=void 0,t=a.shift(),d(t);p()}catch(t){n.error(t)}var t})))};return t.subscribe(Et(n,h,(function(){f=!0,p()}))),function(){}}(n,e,t,r)})))}var Ct=Array.isArray;function Pt(t){return Tt((function(n){return function(t,n){return Ct(n)?t.apply(void 0,x([],w(n))):t(n)}(t,n)}))}var jt=["addListener","removeListener"],Ft=["addEventListener","removeEventListener"],Bt=["on","off"];function Lt(t,n){return function(r){return function(e){return t[r](n,e)}}}var Mt=A((function(t){return function(){t(this),this.name="ObjectUnsubscribedError",this.message="object unsubscribed"}})),qt=function(t){function n(){var n=t.call(this)||this;return n.closed=!1,n.observers=[],n.isStopped=!1,n.hasError=!1,n.thrownError=null,n}return y(n,t),n.prototype.lift=function(t){var n=new zt(this,this);return n.operator=t,n},n.prototype._throwIfClosed=function(){if(this.closed)throw new Mt},n.prototype.next=function(t){var n=this;Q((function(){var r,e;if(n._throwIfClosed(),!n.isStopped){var o=n.observers.slice();try{for(var i=m(o),u=i.next();!u.done;u=i.next())u.value.next(t)}catch(t){r={error:t}}finally{try{u&&!u.done&&(e=i.return)&&e.call(i)}finally{if(r)throw r.error}}}}))},n.prototype.error=function(t){var n=this;Q((function(){if(n._throwIfClosed(),!n.isStopped){n.hasError=n.isStopped=!0,n.thrownError=t;for(var r=n.observers;r.length;)r.shift().error(t)}}))},n.prototype.complete=function(){var t=this;Q((function(){if(t._throwIfClosed(),!t.isStopped){t.isStopped=!0;for(var n=t.observers;n.length;)n.shift().complete()}}))},n.prototype.unsubscribe=function(){this.isStopped=this.closed=!0,this.observers=null},Object.defineProperty(n.prototype,"observed",{get:function(){var t;return(null===(t=this.observers)||void 0===t?void 0:t.length)>0},enumerable:!1,configurable:!0}),n.prototype._trySubscribe=function(n){return this._throwIfClosed(),t.prototype._trySubscribe.call(this,n)},n.prototype._subscribe=function(t){return this._throwIfClosed(),this._checkFinalizedStatuses(t),this._innerSubscribe(t)},n.prototype._innerSubscribe=function(t){var n=this,r=n.hasError,e=n.isStopped,o=n.observers;return r||e?O:(o.push(t),new T((function(){return I(o,t)})))},n.prototype._checkFinalizedStatuses=function(t){var n=this,r=n.hasError,e=n.thrownError,o=n.isStopped;r?t.error(e):o&&t.complete()},n.prototype.asObservable=function(){var t=new st;return t.source=this,t},n.create=function(t,n){return new zt(t,n)},n}(st),zt=function(t){function n(n,r){var e=t.call(this)||this;return e.destination=n,e.source=r,e}return y(n,t),n.prototype.next=function(t){var n,r;null===(r=null===(n=this.destination)||void 0===n?void 0:n.next)||void 0===r||r.call(n,t)},n.prototype.error=function(t){var n,r;null===(r=null===(n=this.destination)||void 0===n?void 0:n.error)||void 0===r||r.call(n,t)},n.prototype.complete=function(){var t,n;null===(n=null===(t=this.destination)||void 0===t?void 0:t.complete)||void 0===n||n.call(t)},n.prototype._subscribe=function(t){var n,r;return null!==(r=null===(n=this.source)||void 0===n?void 0:n.subscribe(t))&&void 0!==r?r:O},n}(qt),Rt=function(t){function n(n){var r=t.call(this)||this;return r._value=n,r}return y(n,t),Object.defineProperty(n.prototype,"value",{get:function(){return this.getValue()},enumerable:!1,configurable:!0}),n.prototype._subscribe=function(n){var r=t.prototype._subscribe.call(this,n);return!r.closed&&n.next(this._value),r},n.prototype.getValue=function(){var t=this,n=t.hasError,r=t.thrownError,e=t._value;if(n)throw r;return this._throwIfClosed(),e},n.prototype.next=function(n){t.prototype.next.call(this,this._value=n)},n}(qt);function Ut(t,n){return void 0===n&&(n=0),At((function(r,e){r.subscribe(Et(e,(function(r){return Ot(e,t,(function(){return e.next(r)}),n)}),(function(){return Ot(e,t,(function(){return e.complete()}),n)}),(function(r){return Ot(e,t,(function(){return e.error(r)}),n)})))}))}function Nt(t,n){return void 0===n&&(n=0),At((function(r,e){e.add(t.schedule((function(){return r.subscribe(e)}),n))}))}function Zt(t,n){if(!t)throw new Error("Iterable cannot be null");return new st((function(r){Ot(r,n,(function(){var e=t[Symbol.asyncIterator]();Ot(r,n,(function(){e.next().then((function(t){t.done?r.complete():r.next(t.value)}))}),0,!0)}))}))}function Gt(t,n){return n?function(t,n){if(null!=t){if(vt(t))return function(t,n){return _t(t).pipe(Nt(n),Ut(n))}(t,n);if(ht(t))return function(t,n){return new st((function(r){var e=0;return n.schedule((function(){e===t.length?r.complete():(r.next(t[e++]),r.closed||this.schedule())}))}))}(t,n);if(dt(t))return function(t,n){return _t(t).pipe(Nt(n),Ut(n))}(t,n);if(yt(t))return Zt(t,n);if(wt(t))return function(t,n){return new st((function(r){var e;return Ot(r,n,(function(){e=t[mt](),Ot(r,n,(function(){var t,n,o;try{n=(t=e.next()).value,o=t.done}catch(t){return void r.error(t)}o?r.complete():r.next(n)}),0,!0)})),function(){return S(null==e?void 0:e.return)&&e.return()}}))}(t,n);if(gt(t))return function(t,n){return Zt(xt(t),n)}(t,n)}throw bt(t)}(t,n):_t(t)}function Yt(t,n,r){return void 0===r&&(r=ut),function(e){Dt(n,(function(){for(var o=t.length,i=new Array(o),u=o,c=o,a=function(o){Dt(n,(function(){var a=Gt(t[o],n),s=!1;a.subscribe(Et(e,(function(t){i[o]=t,s||(s=!0,c--),c||e.next(r(i.slice()))}),(function(){--u||e.complete()})))}),e)},s=0;s<o;s++)a(s)}),e)}}function Dt(t,n,r){t?Ot(r,t,n):n()}var Ht=Array.isArray;function Vt(t){return 1===t.length&&Ht(t[0])?t[0]:t}function Jt(t){return t[t.length-1]}function Kt(t){return S(Jt(t))?t.pop():void 0}function Qt(t){return ft(Jt(t))?t.pop():void 0}function Wt(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];var r=Kt(t);return r?ct(Wt.apply(void 0,x([],w(t))),Pt(r)):At((function(n,r){Yt(x([n],w(Vt(t))))(r)}))}var Xt={schedule:function(t){var n=requestAnimationFrame,r=cancelAnimationFrame,e=Xt.delegate;e&&(n=e.requestAnimationFrame,r=e.cancelAnimationFrame);var o=n((function(n){r=void 0,t(n)}));return new T((function(){return null==r?void 0:r(o)}))},requestAnimationFrame:function(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];var r=Xt.delegate;return((null==r?void 0:r.requestAnimationFrame)||requestAnimationFrame).apply(void 0,x([],w(t)))},cancelAnimationFrame:function(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];var r=Xt.delegate;return((null==r?void 0:r.cancelAnimationFrame)||cancelAnimationFrame).apply(void 0,x([],w(t)))},delegate:void 0},$t=function(t){function n(n,r){var e=t.call(this,n,r)||this;return e.scheduler=n,e.work=r,e}return y(n,t),n.prototype.requestAsyncId=function(n,r,e){return void 0===e&&(e=0),null!==e&&e>0?t.prototype.requestAsyncId.call(this,n,r,e):(n.actions.push(this),n._scheduled||(n._scheduled=Xt.requestAnimationFrame((function(){return n.flush(void 0)}))))},n.prototype.recycleAsyncId=function(n,r,e){if(void 0===e&&(e=0),null!=e&&e>0||null==e&&this.delay>0)return t.prototype.recycleAsyncId.call(this,n,r,e);n.actions.some((function(t){return t.id===r}))||(Xt.cancelAnimationFrame(r),n._scheduled=void 0)},n}(F),tn=function(t){function n(){return null!==t&&t.apply(this,arguments)||this}return y(n,t),n.prototype.flush=function(t){this._active=!0;var n=this._scheduled;this._scheduled=void 0;var r,e=this.actions;t=t||e.shift();do{if(r=t.execute(t.state,t.delay))break}while((t=e[0])&&t.id===n&&e.shift());if(this._active=!1,r){for(;(t=e[0])&&t.id===n&&e.shift();)t.unsubscribe();throw r}},n}(M),nn=new tn($t);function rn(t,n){return At((function(r,e){var o=0;r.subscribe(Et(e,(function(r){return t.call(n,r,o++)&&e.next(r)})))}))}function en(){return void 0===(t=1)&&(t=1/0),kt(ut,t);var t}function on(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];return en()(Gt(t,Qt(t)))}function un(t,n,r,e,o){return function(i,u){var c=r,a=n,s=0;i.subscribe(Et(u,(function(n){var r=s++;a=c?t(a,n,r):(c=!0,n),e&&u.next(a)}),o&&function(){c&&u.next(a),u.complete()}))}}function cn(t,n){return At(un(t,n,arguments.length>=2,!0))}function an(t,n){return void 0===n&&(n=ut),t=null!=t?t:sn,At((function(r,e){var o,i=!0;r.subscribe(Et(e,(function(r){var u=n(r);!i&&t(o,u)||(i=!1,o=u,e.next(r))})))}))}function sn(t,n){return t===n}function ln(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];var r=Kt(t);return At((function(n,e){for(var o=t.length,i=new Array(o),u=t.map((function(){return!1})),c=!1,a=function(n){_t(t[n]).subscribe(Et(e,(function(t){i[n]=t,c||u[n]||(u[n]=!0,(c=u.every(ut))&&(u=null))}),H))},s=0;s<o;s++)a(s);n.subscribe(Et(e,(function(t){if(c){var n=x([t],w(i));e.next(r?r.apply(void 0,x([],w(n))):n)}})))}))}var fn=new st((function(t){return t.complete()}));function pn(t){return t<=0?function(){return fn}:At((function(n,r){var e=0;n.subscribe(Et(r,(function(n){++e<=t&&(r.next(n),t<=e&&r.complete())})))}))}function hn(t,n){for(var r=[],e=2;e<arguments.length;e++)r[e-2]=arguments[e];return!0===n?(t(),null):!1===n?null:n.apply(void 0,x([],w(r))).pipe(pn(1)).subscribe((function(){return t()}))}function dn(t,n){return void 0===n&&(n=!1),At((function(r,e){var o=0;r.subscribe(Et(e,(function(r){var i=t(r,o++);(i||n)&&e.next(r),!i&&e.complete()})))}))}function vn(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];var r=Qt(t);return Gt(t,r)}var yn,bn,mn,wn,xn,gn={a:{x:-1,y:0},d:{x:1,y:0},w:{x:0,y:-1},s:{x:0,y:1}},_n=gn.d,Sn=function(t,n){return t.x===n.x&&t.y===n.y},An=function(t,n){return Math.floor(Math.random()*(n-t+1)+t)},En=function(){return{x:An(0,19),y:An(0,19)}},In=pt(300),Tn=function t(n,r,e,o){if(S(e)&&(o=e,e=void 0),o)return t(n,r,e).pipe(Pt(o));var i=w(function(t){return S(t.addEventListener)&&S(t.removeEventListener)}(n)?Ft.map((function(t){return function(o){return n[t](r,o,e)}})):function(t){return S(t.addListener)&&S(t.removeListener)}(n)?jt.map(Lt(n,r)):function(t){return S(t.on)&&S(t.off)}(n)?Bt.map(Lt(n,r)):[],2),u=i[0],c=i[1];if(!u&&ht(n))return kt((function(n){return t(n,r,e)}))(_t(n));if(!u)throw new TypeError("Invalid event target");return new st((function(t){var n=function(){for(var n=[],r=0;r<arguments.length;r++)n[r]=arguments[r];return t.next(1<n.length?n:n[0])};return u(n),function(){return c(n)}}))}(document,"keydown").pipe((yn=In,At((function(t,n){var r=[];return t.subscribe(Et(n,(function(t){return r.push(t)}),(function(){n.next(r),n.complete()}))),yn.subscribe(Et(n,(function(){var t=r;r=[],n.next(t)}),H)),function(){r=null}}))),rn((function(t){return t.length>0})),Tt((function(t){return gn[t[t.length-1].key]})),rn((function(t){return Boolean(t)})),function(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];var r=Qt(t);return At((function(n,e){(r?on(t,n,r):on(t,n)).subscribe(e)}))}(_n),cn((function(t,n){return e=n,(r=t).x===-1*e.x||r.y===-1*e.y?t:n;var r,e})),an()),On=new Rt(0),kn=On.pipe(cn((function(t,n){return t+n}),2)),Cn=In.pipe(ln(Tn,kn,(function(t,n,r){return[n,r]})),cn((function(t,n){var r=n[0],e=n[1],o=r,i=function(t,n,r){if(r||2===arguments.length)for(var e,o=0,i=n.length;o<i;o++)!e&&o in n||(e||(e=Array.prototype.slice.call(n,0,o)),e[o]=n[o]);return t.concat(e||Array.prototype.slice.call(n))}([{x:t[0].x+o.x,y:t[0].y+o.y}],t,!0);return e<i.length&&i.pop(),i}),function(t){for(var n=[],r=1;r>=0;r--)n.push({x:r,y:0});return n}())),Pn=Cn.pipe(cn((function(t,n){for(var r=t,e=n.some((function(t){return Sn(t,r)}));e;)r=En(),e=n.some((function(t){return Sn(t,r)}));return r}),En()),an(),function(t){void 0===t&&(t={});var n=t.connector,r=void 0===n?function(){return new qt}:n,e=t.resetOnError,o=void 0===e||e,i=t.resetOnComplete,u=void 0===i||i,c=t.resetOnRefCountZero,a=void 0===c||c;return function(t){var n=null,e=null,i=null,c=0,s=!1,l=!1,f=function(){null==e||e.unsubscribe(),e=null},p=function(){f(),n=i=null,s=l=!1},h=function(){var t=n;p(),null==t||t.unsubscribe()};return At((function(t,d){c++,l||s||f();var v=i=null!=i?i:r();d.add((function(){0!=--c||l||s||(e=hn(h,a))})),v.subscribe(d),n||(n=new nt({next:function(t){return v.next(t)},error:function(t){l=!0,f(),e=hn(p,o,t),v.error(t)},complete:function(){s=!0,f(),e=hn(p,u),v.complete()}}),Gt(t).subscribe(n))}))(t)}}());Pn.pipe((1,rn((function(t,n){return 1<=n}))),(bn=function(){On.next(1)},xn=S(bn)?{next:bn,error:mn,complete:wn}:bn,xn?At((function(t,n){var r;null===(r=xn.subscribe)||void 0===r||r.call(xn);var e=!0;t.subscribe(Et(n,(function(t){var r;null===(r=xn.next)||void 0===r||r.call(xn,t),n.next(t)}),(function(){var t;e=!1,null===(t=xn.complete)||void 0===t||t.call(xn),n.complete()}),(function(t){var r;e=!1,null===(r=xn.error)||void 0===r||r.call(xn,t),n.error(t)}),(function(){var t,n;e&&(null===(t=xn.unsubscribe)||void 0===t||t.call(xn)),null===(n=xn.finalize)||void 0===n||n.call(xn)})))})):ut)).subscribe();var jn=Cn.pipe(function(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];return Wt.apply(void 0,x([],w(t)))}(Pn),dn((function(t){var n=t[0],r=n[0];return!(r.x<0||r.y<0||r.x>19||r.y>19||function(t,n){for(var r=!1,e=3;e<n.length;e+=1)Sn(t,n[e])&&(r=!0);return r}(r,n))})),function(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];return function(n){return on(n,vn.apply(void 0,x([],w(t))))}}("end")),Fn=pt(100,nn).pipe(ln(jn,(function(t,n){return n})),dn((function(t){return"end"!==t}))),Bn=function(){function t(t){var n=this;this.renderGameOver=function(){n.ctx.font="bold 15px sans-serif",n.ctx.fillStyle="black",n.ctx.textAlign="center",n.ctx.textBaseline="middle",n.ctx.fillText("Game Over",150,90)},this.ctx=t}return t.prototype.getSnakeBoxColor=function(t){return 0===t?"#3f72af":"#c7d0d5"},t.prototype.drawBox=function(t,n){this.ctx.fillStyle=n,this.ctx.fillRect(15*t.x,15*t.y,15,15)},t.prototype.drawBackground=function(){this.ctx.fillStyle="#EEE",this.ctx.fillRect(0,0,300,300)},t.prototype.drawSnake=function(t){var n=this;t.forEach((function(t,r){n.drawBox(t,n.getSnakeBoxColor(r))}))},t.prototype.drawApple=function(t){this.drawBox(t,"#ff7148")},t.prototype.drawScore=function(t){this.ctx.font="bold 80px sans-serif",this.ctx.fillStyle="#d9d9d9",this.ctx.textAlign="center",this.ctx.textBaseline="middle",this.ctx.fillText(t.toString(),150,150)},t.prepareCanvasSize=function(t){t.width=300,t.height=300},t}();const Ln=Bn;var Mn=document.getElementById("appCan"),qn=new Ln(Mn.getContext("2d"));Ln.prepareCanvasSize(Mn),Fn.subscribe({next:function(t){var n=t[0],r=t[1];qn.drawBackground(),qn.drawScore(n.length-2),qn.drawSnake(n),qn.drawApple(r)},complete:function(){return qn.renderGameOver()}})})()})();
+(()=>{"use strict";var t={935:(t,n,r)=>{r.d(n,{Z:()=>c});var e=r(81),o=r.n(e),i=r(645),u=r.n(i)()(o());u.push([t.id,'h1,\nh2,\nh3,\nh4,\nh5,\nspan,\na,\np,\nbutton {\n  font-family: "Play", sans-serif;\n}\nheader h1 {\n  color: #ff7148;\n  text-align: center;\n}\ncanvas {\n  display: block;\n  border: 5px solid #93b1c6;\n  margin: 20px auto;\n  width: 400px;\n  height: 400px;\n}\n.gameControl {\n  text-align: center;\n  display: none;\n}\n#gameStartBtn {\n  padding: 8px 50px;\n  font-size: 1.2em;\n}\n#gameStartBtn:focus {\n  outline: none;\n}\n',""]);const c=u},645:t=>{t.exports=function(t){var n=[];return n.toString=function(){return this.map((function(n){var r="",e=void 0!==n[5];return n[4]&&(r+="@supports (".concat(n[4],") {")),n[2]&&(r+="@media ".concat(n[2]," {")),e&&(r+="@layer".concat(n[5].length>0?" ".concat(n[5]):""," {")),r+=t(n),e&&(r+="}"),n[2]&&(r+="}"),n[4]&&(r+="}"),r})).join("")},n.i=function(t,r,e,o,i){"string"==typeof t&&(t=[[null,t,void 0]]);var u={};if(e)for(var c=0;c<this.length;c++){var a=this[c][0];null!=a&&(u[a]=!0)}for(var s=0;s<t.length;s++){var l=[].concat(t[s]);e&&u[l[0]]||(void 0!==i&&(void 0===l[5]||(l[1]="@layer".concat(l[5].length>0?" ".concat(l[5]):""," {").concat(l[1],"}")),l[5]=i),r&&(l[2]?(l[1]="@media ".concat(l[2]," {").concat(l[1],"}"),l[2]=r):l[2]=r),o&&(l[4]?(l[1]="@supports (".concat(l[4],") {").concat(l[1],"}"),l[4]=o):l[4]="".concat(o)),n.push(l))}},n}},81:t=>{t.exports=function(t){return t[1]}},379:t=>{var n=[];function r(t){for(var r=-1,e=0;e<n.length;e++)if(n[e].identifier===t){r=e;break}return r}function e(t,e){for(var i={},u=[],c=0;c<t.length;c++){var a=t[c],s=e.base?a[0]+e.base:a[0],l=i[s]||0,f="".concat(s," ").concat(l);i[s]=l+1;var p=r(f),h={css:a[1],media:a[2],sourceMap:a[3],supports:a[4],layer:a[5]};if(-1!==p)n[p].references++,n[p].updater(h);else{var d=o(h,e);e.byIndex=c,n.splice(c,0,{identifier:f,updater:d,references:1})}u.push(f)}return u}function o(t,n){var r=n.domAPI(n);return r.update(t),function(n){if(n){if(n.css===t.css&&n.media===t.media&&n.sourceMap===t.sourceMap&&n.supports===t.supports&&n.layer===t.layer)return;r.update(t=n)}else r.remove()}}t.exports=function(t,o){var i=e(t=t||[],o=o||{});return function(t){t=t||[];for(var u=0;u<i.length;u++){var c=r(i[u]);n[c].references--}for(var a=e(t,o),s=0;s<i.length;s++){var l=r(i[s]);0===n[l].references&&(n[l].updater(),n.splice(l,1))}i=a}}},569:t=>{var n={};t.exports=function(t,r){var e=function(t){if(void 0===n[t]){var r=document.querySelector(t);if(window.HTMLIFrameElement&&r instanceof window.HTMLIFrameElement)try{r=r.contentDocument.head}catch(t){r=null}n[t]=r}return n[t]}(t);if(!e)throw new Error("Couldn't find a style target. This probably means that the value for the 'insert' parameter is invalid.");e.appendChild(r)}},216:t=>{t.exports=function(t){var n=document.createElement("style");return t.setAttributes(n,t.attributes),t.insert(n,t.options),n}},565:(t,n,r)=>{t.exports=function(t){var n=r.nc;n&&t.setAttribute("nonce",n)}},795:t=>{t.exports=function(t){var n=t.insertStyleElement(t);return{update:function(r){!function(t,n,r){var e="";r.supports&&(e+="@supports (".concat(r.supports,") {")),r.media&&(e+="@media ".concat(r.media," {"));var o=void 0!==r.layer;o&&(e+="@layer".concat(r.layer.length>0?" ".concat(r.layer):""," {")),e+=r.css,o&&(e+="}"),r.media&&(e+="}"),r.supports&&(e+="}");var i=r.sourceMap;i&&"undefined"!=typeof btoa&&(e+="\n/*# sourceMappingURL=data:application/json;base64,".concat(btoa(unescape(encodeURIComponent(JSON.stringify(i))))," */")),n.styleTagTransform(e,t,n.options)}(n,t,r)},remove:function(){!function(t){if(null===t.parentNode)return!1;t.parentNode.removeChild(t)}(n)}}}},589:t=>{t.exports=function(t,n){if(n.styleSheet)n.styleSheet.cssText=t;else{for(;n.firstChild;)n.removeChild(n.firstChild);n.appendChild(document.createTextNode(t))}}}},n={};function r(e){var o=n[e];if(void 0!==o)return o.exports;var i=n[e]={id:e,exports:{}};return t[e](i,i.exports,r),i.exports}r.n=t=>{var n=t&&t.__esModule?()=>t.default:()=>t;return r.d(n,{a:n}),n},r.d=(t,n)=>{for(var e in n)r.o(n,e)&&!r.o(t,e)&&Object.defineProperty(t,e,{enumerable:!0,get:n[e]})},r.o=(t,n)=>Object.prototype.hasOwnProperty.call(t,n),r.nc=void 0,(()=>{var t=r(379),n=r.n(t),e=r(795),o=r.n(e),i=r(569),u=r.n(i),c=r(565),a=r.n(c),s=r(216),l=r.n(s),f=r(589),p=r.n(f),h=r(935),d={};d.styleTagTransform=p(),d.setAttributes=a(),d.insert=u().bind(null,"head"),d.domAPI=o(),d.insertStyleElement=l(),n()(h.Z,d),h.Z&&h.Z.locals&&h.Z.locals;var v=function(t,n){return v=Object.setPrototypeOf||{__proto__:[]}instanceof Array&&function(t,n){t.__proto__=n}||function(t,n){for(var r in n)Object.prototype.hasOwnProperty.call(n,r)&&(t[r]=n[r])},v(t,n)};function y(t,n){if("function"!=typeof n&&null!==n)throw new TypeError("Class extends value "+String(n)+" is not a constructor or null");function r(){this.constructor=t}v(t,n),t.prototype=null===n?Object.create(n):(r.prototype=n.prototype,new r)}function b(t,n){var r,e,o,i,u={label:0,sent:function(){if(1&o[0])throw o[1];return o[1]},trys:[],ops:[]};return i={next:c(0),throw:c(1),return:c(2)},"function"==typeof Symbol&&(i[Symbol.iterator]=function(){return this}),i;function c(c){return function(a){return function(c){if(r)throw new TypeError("Generator is already executing.");for(;i&&(i=0,c[0]&&(u=0)),u;)try{if(r=1,e&&(o=2&c[0]?e.return:c[0]?e.throw||((o=e.return)&&o.call(e),0):e.next)&&!(o=o.call(e,c[1])).done)return o;switch(e=0,o&&(c=[2&c[0],o.value]),c[0]){case 0:case 1:o=c;break;case 4:return u.label++,{value:c[1],done:!1};case 5:u.label++,e=c[1],c=[0];continue;case 7:c=u.ops.pop(),u.trys.pop();continue;default:if(!((o=(o=u.trys).length>0&&o[o.length-1])||6!==c[0]&&2!==c[0])){u=0;continue}if(3===c[0]&&(!o||c[1]>o[0]&&c[1]<o[3])){u.label=c[1];break}if(6===c[0]&&u.label<o[1]){u.label=o[1],o=c;break}if(o&&u.label<o[2]){u.label=o[2],u.ops.push(c);break}o[2]&&u.ops.pop(),u.trys.pop();continue}c=n.call(t,u)}catch(t){c=[6,t],e=0}finally{r=o=0}if(5&c[0])throw c[1];return{value:c[0]?c[1]:void 0,done:!0}}([c,a])}}}function m(t){var n="function"==typeof Symbol&&Symbol.iterator,r=n&&t[n],e=0;if(r)return r.call(t);if(t&&"number"==typeof t.length)return{next:function(){return t&&e>=t.length&&(t=void 0),{value:t&&t[e++],done:!t}}};throw new TypeError(n?"Object is not iterable.":"Symbol.iterator is not defined.")}function x(t,n){var r="function"==typeof Symbol&&t[Symbol.iterator];if(!r)return t;var e,o,i=r.call(t),u=[];try{for(;(void 0===n||n-- >0)&&!(e=i.next()).done;)u.push(e.value)}catch(t){o={error:t}}finally{try{e&&!e.done&&(r=i.return)&&r.call(i)}finally{if(o)throw o.error}}return u}function w(t,n,r){if(r||2===arguments.length)for(var e,o=0,i=n.length;o<i;o++)!e&&o in n||(e||(e=Array.prototype.slice.call(n,0,o)),e[o]=n[o]);return t.concat(e||Array.prototype.slice.call(n))}function g(t){return this instanceof g?(this.v=t,this):new g(t)}function _(t,n,r){if(!Symbol.asyncIterator)throw new TypeError("Symbol.asyncIterator is not defined.");var e,o=r.apply(t,n||[]),i=[];return e={},u("next"),u("throw"),u("return"),e[Symbol.asyncIterator]=function(){return this},e;function u(t){o[t]&&(e[t]=function(n){return new Promise((function(r,e){i.push([t,n,r,e])>1||c(t,n)}))})}function c(t,n){try{(r=o[t](n)).value instanceof g?Promise.resolve(r.value.v).then(a,s):l(i[0][2],r)}catch(t){l(i[0][3],t)}var r}function a(t){c("next",t)}function s(t){c("throw",t)}function l(t,n){t(n),i.shift(),i.length&&c(i[0][0],i[0][1])}}function S(t){return"function"==typeof t}function A(t){var n=t((function(t){Error.call(t),t.stack=(new Error).stack}));return n.prototype=Object.create(Error.prototype),n.prototype.constructor=n,n}Object.create,Object.create;var I=A((function(t){return function(n){t(this),this.message=n?n.length+" errors occurred during unsubscription:\n"+n.map((function(t,n){return n+1+") "+t.toString()})).join("\n  "):"",this.name="UnsubscriptionError",this.errors=n}}));function E(t,n){if(t){var r=t.indexOf(n);0<=r&&t.splice(r,1)}}var O=function(){function t(t){this.initialTeardown=t,this.closed=!1,this._parentage=null,this._finalizers=null}var n;return t.prototype.unsubscribe=function(){var t,n,r,e,o;if(!this.closed){this.closed=!0;var i=this._parentage;if(i)if(this._parentage=null,Array.isArray(i))try{for(var u=m(i),c=u.next();!c.done;c=u.next())c.value.remove(this)}catch(n){t={error:n}}finally{try{c&&!c.done&&(n=u.return)&&n.call(u)}finally{if(t)throw t.error}}else i.remove(this);var a=this.initialTeardown;if(S(a))try{a()}catch(t){o=t instanceof I?t.errors:[t]}var s=this._finalizers;if(s){this._finalizers=null;try{for(var l=m(s),f=l.next();!f.done;f=l.next()){var p=f.value;try{C(p)}catch(t){o=null!=o?o:[],t instanceof I?o=w(w([],x(o)),x(t.errors)):o.push(t)}}}catch(t){r={error:t}}finally{try{f&&!f.done&&(e=l.return)&&e.call(l)}finally{if(r)throw r.error}}}if(o)throw new I(o)}},t.prototype.add=function(n){var r;if(n&&n!==this)if(this.closed)C(n);else{if(n instanceof t){if(n.closed||n._hasParent(this))return;n._addParent(this)}(this._finalizers=null!==(r=this._finalizers)&&void 0!==r?r:[]).push(n)}},t.prototype._hasParent=function(t){var n=this._parentage;return n===t||Array.isArray(n)&&n.includes(t)},t.prototype._addParent=function(t){var n=this._parentage;this._parentage=Array.isArray(n)?(n.push(t),n):n?[n,t]:t},t.prototype._removeParent=function(t){var n=this._parentage;n===t?this._parentage=null:Array.isArray(n)&&E(n,t)},t.prototype.remove=function(n){var r=this._finalizers;r&&E(r,n),n instanceof t&&n._removeParent(this)},t.EMPTY=((n=new t).closed=!0,n),t}(),T=O.EMPTY;function k(t){return t instanceof O||t&&"closed"in t&&S(t.remove)&&S(t.add)&&S(t.unsubscribe)}function C(t){S(t)?t():t.unsubscribe()}var P=function(t){function n(n,r){return t.call(this)||this}return y(n,t),n.prototype.schedule=function(t,n){return void 0===n&&(n=0),this},n}(O),j={setInterval:function(t,n){for(var r=[],e=2;e<arguments.length;e++)r[e-2]=arguments[e];var o=j.delegate;return(null==o?void 0:o.setInterval)?o.setInterval.apply(o,w([t,n],x(r))):setInterval.apply(void 0,w([t,n],x(r)))},clearInterval:function(t){var n=j.delegate;return((null==n?void 0:n.clearInterval)||clearInterval)(t)},delegate:void 0},F=function(t){function n(n,r){var e=t.call(this,n,r)||this;return e.scheduler=n,e.work=r,e.pending=!1,e}return y(n,t),n.prototype.schedule=function(t,n){var r;if(void 0===n&&(n=0),this.closed)return this;this.state=t;var e=this.id,o=this.scheduler;return null!=e&&(this.id=this.recycleAsyncId(o,e,n)),this.pending=!0,this.delay=n,this.id=null!==(r=this.id)&&void 0!==r?r:this.requestAsyncId(o,this.id,n),this},n.prototype.requestAsyncId=function(t,n,r){return void 0===r&&(r=0),j.setInterval(t.flush.bind(t,this),r)},n.prototype.recycleAsyncId=function(t,n,r){if(void 0===r&&(r=0),null!=r&&this.delay===r&&!1===this.pending)return n;null!=n&&j.clearInterval(n)},n.prototype.execute=function(t,n){if(this.closed)return new Error("executing a cancelled action");this.pending=!1;var r=this._execute(t,n);if(r)return r;!1===this.pending&&null!=this.id&&(this.id=this.recycleAsyncId(this.scheduler,this.id,null))},n.prototype._execute=function(t,n){var r,e=!1;try{this.work(t)}catch(t){e=!0,r=t||new Error("Scheduled action threw falsy error")}if(e)return this.unsubscribe(),r},n.prototype.unsubscribe=function(){if(!this.closed){var n=this.id,r=this.scheduler,e=r.actions;this.work=this.state=this.scheduler=null,this.pending=!1,E(e,this),null!=n&&(this.id=this.recycleAsyncId(r,n,null)),this.delay=null,t.prototype.unsubscribe.call(this)}},n}(P),z={now:function(){return(z.delegate||Date).now()},delegate:void 0},B=function(){function t(n,r){void 0===r&&(r=t.now),this.schedulerActionCtor=n,this.now=r}return t.prototype.schedule=function(t,n,r){return void 0===n&&(n=0),new this.schedulerActionCtor(this,t).schedule(r,n)},t.now=z.now,t}(),L=function(t){function n(n,r){void 0===r&&(r=B.now);var e=t.call(this,n,r)||this;return e.actions=[],e._active=!1,e}return y(n,t),n.prototype.flush=function(t){var n=this.actions;if(this._active)n.push(t);else{var r;this._active=!0;do{if(r=t.execute(t.state,t.delay))break}while(t=n.shift());if(this._active=!1,r){for(;t=n.shift();)t.unsubscribe();throw r}}},n}(B),M=new L(F),q=M,R=null,U=null,N=void 0,Z=!1,G=!1,Y={setTimeout:function(t,n){for(var r=[],e=2;e<arguments.length;e++)r[e-2]=arguments[e];var o=Y.delegate;return(null==o?void 0:o.setTimeout)?o.setTimeout.apply(o,w([t,n],x(r))):setTimeout.apply(void 0,w([t,n],x(r)))},clearTimeout:function(t){var n=Y.delegate;return((null==n?void 0:n.clearTimeout)||clearTimeout)(t)},delegate:void 0};function D(t){Y.setTimeout((function(){if(!R)throw t;R(t)}))}function H(){}var V=J("C",void 0,void 0);function J(t,n,r){return{kind:t,value:n,error:r}}var K=null;function Q(t){if(Z){var n=!K;if(n&&(K={errorThrown:!1,error:null}),t(),n){var r=K,e=r.errorThrown,o=r.error;if(K=null,e)throw o}}else t()}var W=function(t){function n(n){var r=t.call(this)||this;return r.isStopped=!1,n?(r.destination=n,k(n)&&n.add(r)):r.destination=ot,r}return y(n,t),n.create=function(t,n,r){return new nt(t,n,r)},n.prototype.next=function(t){this.isStopped?et(function(t){return J("N",t,void 0)}(t),this):this._next(t)},n.prototype.error=function(t){this.isStopped?et(J("E",void 0,t),this):(this.isStopped=!0,this._error(t))},n.prototype.complete=function(){this.isStopped?et(V,this):(this.isStopped=!0,this._complete())},n.prototype.unsubscribe=function(){this.closed||(this.isStopped=!0,t.prototype.unsubscribe.call(this),this.destination=null)},n.prototype._next=function(t){this.destination.next(t)},n.prototype._error=function(t){try{this.destination.error(t)}finally{this.unsubscribe()}},n.prototype._complete=function(){try{this.destination.complete()}finally{this.unsubscribe()}},n}(O),X=Function.prototype.bind;function $(t,n){return X.call(t,n)}var tt=function(){function t(t){this.partialObserver=t}return t.prototype.next=function(t){var n=this.partialObserver;if(n.next)try{n.next(t)}catch(t){rt(t)}},t.prototype.error=function(t){var n=this.partialObserver;if(n.error)try{n.error(t)}catch(t){rt(t)}else rt(t)},t.prototype.complete=function(){var t=this.partialObserver;if(t.complete)try{t.complete()}catch(t){rt(t)}},t}(),nt=function(t){function n(n,r,e){var o,i,u=t.call(this)||this;return S(n)||!n?o={next:null!=n?n:void 0,error:null!=r?r:void 0,complete:null!=e?e:void 0}:u&&G?((i=Object.create(n)).unsubscribe=function(){return u.unsubscribe()},o={next:n.next&&$(n.next,i),error:n.error&&$(n.error,i),complete:n.complete&&$(n.complete,i)}):o=n,u.destination=new tt(o),u}return y(n,t),n}(W);function rt(t){var n;Z?(n=t,Z&&K&&(K.errorThrown=!0,K.error=n)):D(t)}function et(t,n){var r=U;r&&Y.setTimeout((function(){return r(t,n)}))}var ot={closed:!0,next:H,error:function(t){throw t},complete:H},it="function"==typeof Symbol&&Symbol.observable||"@@observable";function ut(t){return t}function ct(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];return at(t)}function at(t){return 0===t.length?ut:1===t.length?t[0]:function(n){return t.reduce((function(t,n){return n(t)}),n)}}var st=function(){function t(t){t&&(this._subscribe=t)}return t.prototype.lift=function(n){var r=new t;return r.source=this,r.operator=n,r},t.prototype.subscribe=function(t,n,r){var e,o=this,i=(e=t)&&e instanceof W||function(t){return t&&S(t.next)&&S(t.error)&&S(t.complete)}(e)&&k(e)?t:new nt(t,n,r);return Q((function(){var t=o,n=t.operator,r=t.source;i.add(n?n.call(i,r):r?o._subscribe(i):o._trySubscribe(i))})),i},t.prototype._trySubscribe=function(t){try{return this._subscribe(t)}catch(n){t.error(n)}},t.prototype.forEach=function(t,n){var r=this;return new(n=lt(n))((function(n,e){var o=new nt({next:function(n){try{t(n)}catch(t){e(t),o.unsubscribe()}},error:e,complete:n});r.subscribe(o)}))},t.prototype._subscribe=function(t){var n;return null===(n=this.source)||void 0===n?void 0:n.subscribe(t)},t.prototype[it]=function(){return this},t.prototype.pipe=function(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];return at(t)(this)},t.prototype.toPromise=function(t){var n=this;return new(t=lt(t))((function(t,r){var e;n.subscribe((function(t){return e=t}),(function(t){return r(t)}),(function(){return t(e)}))}))},t.create=function(n){return new t(n)},t}();function lt(t){var n;return null!==(n=null!=t?t:N)&&void 0!==n?n:Promise}function ft(t){return t&&S(t.schedule)}function pt(t,n){return void 0===t&&(t=0),void 0===n&&(n=M),t<0&&(t=0),function(t,n,r){void 0===t&&(t=0),void 0===r&&(r=q);var e=-1;return null!=n&&(ft(n)?r=n:e=n),new st((function(n){var o,i=(o=t)instanceof Date&&!isNaN(o)?+t-r.now():t;i<0&&(i=0);var u=0;return r.schedule((function(){n.closed||(n.next(u++),0<=e?this.schedule(void 0,e):n.complete())}),i)}))}(t,t,n)}var ht=function(t){return t&&"number"==typeof t.length&&"function"!=typeof t};function dt(t){return S(null==t?void 0:t.then)}function vt(t){return S(t[it])}function yt(t){return Symbol.asyncIterator&&S(null==t?void 0:t[Symbol.asyncIterator])}function bt(t){return new TypeError("You provided "+(null!==t&&"object"==typeof t?"an invalid object":"'"+t+"'")+" where a stream was expected. You can provide an Observable, Promise, ReadableStream, Array, AsyncIterable, or Iterable.")}var mt="function"==typeof Symbol&&Symbol.iterator?Symbol.iterator:"@@iterator";function xt(t){return S(null==t?void 0:t[mt])}function wt(t){return _(this,arguments,(function(){var n,r,e;return b(this,(function(o){switch(o.label){case 0:n=t.getReader(),o.label=1;case 1:o.trys.push([1,,9,10]),o.label=2;case 2:return[4,g(n.read())];case 3:return r=o.sent(),e=r.value,r.done?[4,g(void 0)]:[3,5];case 4:return[2,o.sent()];case 5:return[4,g(e)];case 6:return[4,o.sent()];case 7:return o.sent(),[3,2];case 8:return[3,10];case 9:return n.releaseLock(),[7];case 10:return[2]}}))}))}function gt(t){return S(null==t?void 0:t.getReader)}function _t(t){if(t instanceof st)return t;if(null!=t){if(vt(t))return o=t,new st((function(t){var n=o[it]();if(S(n.subscribe))return n.subscribe(t);throw new TypeError("Provided object does not correctly implement Symbol.observable")}));if(ht(t))return e=t,new st((function(t){for(var n=0;n<e.length&&!t.closed;n++)t.next(e[n]);t.complete()}));if(dt(t))return r=t,new st((function(t){r.then((function(n){t.closed||(t.next(n),t.complete())}),(function(n){return t.error(n)})).then(null,D)}));if(yt(t))return St(t);if(xt(t))return n=t,new st((function(t){var r,e;try{for(var o=m(n),i=o.next();!i.done;i=o.next()){var u=i.value;if(t.next(u),t.closed)return}}catch(t){r={error:t}}finally{try{i&&!i.done&&(e=o.return)&&e.call(o)}finally{if(r)throw r.error}}t.complete()}));if(gt(t))return St(wt(t))}var n,r,e,o;throw bt(t)}function St(t){return new st((function(n){(function(t,n){var r,e,o,i,u,c,a,s;return u=this,c=void 0,s=function(){var u,c;return b(this,(function(a){switch(a.label){case 0:a.trys.push([0,5,6,11]),r=function(t){if(!Symbol.asyncIterator)throw new TypeError("Symbol.asyncIterator is not defined.");var n,r=t[Symbol.asyncIterator];return r?r.call(t):(t=m(t),n={},e("next"),e("throw"),e("return"),n[Symbol.asyncIterator]=function(){return this},n);function e(r){n[r]=t[r]&&function(n){return new Promise((function(e,o){!function(t,n,r,e){Promise.resolve(e).then((function(n){t({value:n,done:r})}),n)}(e,o,(n=t[r](n)).done,n.value)}))}}}(t),a.label=1;case 1:return[4,r.next()];case 2:if((e=a.sent()).done)return[3,4];if(u=e.value,n.next(u),n.closed)return[2];a.label=3;case 3:return[3,1];case 4:return[3,11];case 5:return c=a.sent(),o={error:c},[3,11];case 6:return a.trys.push([6,,9,10]),e&&!e.done&&(i=r.return)?[4,i.call(r)]:[3,8];case 7:a.sent(),a.label=8;case 8:return[3,10];case 9:if(o)throw o.error;return[7];case 10:return[7];case 11:return n.complete(),[2]}}))},new((a=void 0)||(a=Promise))((function(t,n){function r(t){try{o(s.next(t))}catch(t){n(t)}}function e(t){try{o(s.throw(t))}catch(t){n(t)}}function o(n){var o;n.done?t(n.value):(o=n.value,o instanceof a?o:new a((function(t){t(o)}))).then(r,e)}o((s=s.apply(u,c||[])).next())}))})(t,n).catch((function(t){return n.error(t)}))}))}function At(t){return function(n){if(function(t){return S(null==t?void 0:t.lift)}(n))return n.lift((function(n){try{return t(n,this)}catch(t){this.error(t)}}));throw new TypeError("Unable to lift unknown Observable type")}}function It(t,n,r,e,o){return new Et(t,n,r,e,o)}var Et=function(t){function n(n,r,e,o,i,u){var c=t.call(this,n)||this;return c.onFinalize=i,c.shouldUnsubscribe=u,c._next=r?function(t){try{r(t)}catch(t){n.error(t)}}:t.prototype._next,c._error=o?function(t){try{o(t)}catch(t){n.error(t)}finally{this.unsubscribe()}}:t.prototype._error,c._complete=e?function(){try{e()}catch(t){n.error(t)}finally{this.unsubscribe()}}:t.prototype._complete,c}return y(n,t),n.prototype.unsubscribe=function(){var n;if(!this.shouldUnsubscribe||this.shouldUnsubscribe()){var r=this.closed;t.prototype.unsubscribe.call(this),!r&&(null===(n=this.onFinalize)||void 0===n||n.call(this))}},n}(W);function Ot(t,n){return At((function(r,e){var o=0;r.subscribe(It(e,(function(r){e.next(t.call(n,r,o++))})))}))}function Tt(t,n,r,e,o){void 0===e&&(e=0),void 0===o&&(o=!1);var i=n.schedule((function(){r(),o?t.add(this.schedule(null,e)):this.unsubscribe()}),e);if(t.add(i),!o)return i}function kt(t,n,r){return void 0===r&&(r=1/0),S(n)?kt((function(r,e){return Ot((function(t,o){return n(r,t,e,o)}))(_t(t(r,e)))}),r):("number"==typeof n&&(r=n),At((function(n,e){return function(t,n,r,e,o,i,u,c){var a=[],s=0,l=0,f=!1,p=function(){!f||a.length||s||n.complete()},h=function(t){return s<e?d(t):a.push(t)},d=function(t){s++;var o=!1;_t(r(t,l++)).subscribe(It(n,(function(t){n.next(t)}),(function(){o=!0}),void 0,(function(){if(o)try{s--;for(;a.length&&s<e;)t=void 0,t=a.shift(),d(t);p()}catch(t){n.error(t)}var t})))};return t.subscribe(It(n,h,(function(){f=!0,p()}))),function(){}}(n,e,t,r)})))}var Ct=Array.isArray;function Pt(t){return Ot((function(n){return function(t,n){return Ct(n)?t.apply(void 0,w([],x(n))):t(n)}(t,n)}))}var jt=["addListener","removeListener"],Ft=["addEventListener","removeEventListener"],zt=["on","off"];function Bt(t,n){return function(r){return function(e){return t[r](n,e)}}}function Lt(t,n){return At((function(r,e){var o=0;r.subscribe(It(e,(function(r){return t.call(n,r,o++)&&e.next(r)})))}))}function Mt(){return void 0===(t=1)&&(t=1/0),kt(ut,t);var t}function qt(t){return t[t.length-1]}function Rt(t){return S(qt(t))?t.pop():void 0}function Ut(t){return ft(qt(t))?t.pop():void 0}function Nt(t,n){return void 0===n&&(n=0),At((function(r,e){r.subscribe(It(e,(function(r){return Tt(e,t,(function(){return e.next(r)}),n)}),(function(){return Tt(e,t,(function(){return e.complete()}),n)}),(function(r){return Tt(e,t,(function(){return e.error(r)}),n)})))}))}function Zt(t,n){return void 0===n&&(n=0),At((function(r,e){e.add(t.schedule((function(){return r.subscribe(e)}),n))}))}function Gt(t,n){if(!t)throw new Error("Iterable cannot be null");return new st((function(r){Tt(r,n,(function(){var e=t[Symbol.asyncIterator]();Tt(r,n,(function(){e.next().then((function(t){t.done?r.complete():r.next(t.value)}))}),0,!0)}))}))}function Yt(t,n){return n?function(t,n){if(null!=t){if(vt(t))return function(t,n){return _t(t).pipe(Zt(n),Nt(n))}(t,n);if(ht(t))return function(t,n){return new st((function(r){var e=0;return n.schedule((function(){e===t.length?r.complete():(r.next(t[e++]),r.closed||this.schedule())}))}))}(t,n);if(dt(t))return function(t,n){return _t(t).pipe(Zt(n),Nt(n))}(t,n);if(yt(t))return Gt(t,n);if(xt(t))return function(t,n){return new st((function(r){var e;return Tt(r,n,(function(){e=t[mt](),Tt(r,n,(function(){var t,n,o;try{n=(t=e.next()).value,o=t.done}catch(t){return void r.error(t)}o?r.complete():r.next(n)}),0,!0)})),function(){return S(null==e?void 0:e.return)&&e.return()}}))}(t,n);if(gt(t))return function(t,n){return Gt(wt(t),n)}(t,n)}throw bt(t)}(t,n):_t(t)}function Dt(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];return Mt()(Yt(t,Ut(t)))}function Ht(t,n,r,e,o){return function(i,u){var c=r,a=n,s=0;i.subscribe(It(u,(function(n){var r=s++;a=c?t(a,n,r):(c=!0,n),e&&u.next(a)}),o&&function(){c&&u.next(a),u.complete()}))}}function Vt(t,n){return At(Ht(t,n,arguments.length>=2,!0))}function Jt(t,n){return void 0===n&&(n=ut),t=null!=t?t:Kt,At((function(r,e){var o,i=!0;r.subscribe(It(e,(function(r){var u=n(r);!i&&t(o,u)||(i=!1,o=u,e.next(r))})))}))}function Kt(t,n){return t===n}var Qt=A((function(t){return function(){t(this),this.name="ObjectUnsubscribedError",this.message="object unsubscribed"}})),Wt=function(t){function n(){var n=t.call(this)||this;return n.closed=!1,n.currentObservers=null,n.observers=[],n.isStopped=!1,n.hasError=!1,n.thrownError=null,n}return y(n,t),n.prototype.lift=function(t){var n=new Xt(this,this);return n.operator=t,n},n.prototype._throwIfClosed=function(){if(this.closed)throw new Qt},n.prototype.next=function(t){var n=this;Q((function(){var r,e;if(n._throwIfClosed(),!n.isStopped){n.currentObservers||(n.currentObservers=Array.from(n.observers));try{for(var o=m(n.currentObservers),i=o.next();!i.done;i=o.next())i.value.next(t)}catch(t){r={error:t}}finally{try{i&&!i.done&&(e=o.return)&&e.call(o)}finally{if(r)throw r.error}}}}))},n.prototype.error=function(t){var n=this;Q((function(){if(n._throwIfClosed(),!n.isStopped){n.hasError=n.isStopped=!0,n.thrownError=t;for(var r=n.observers;r.length;)r.shift().error(t)}}))},n.prototype.complete=function(){var t=this;Q((function(){if(t._throwIfClosed(),!t.isStopped){t.isStopped=!0;for(var n=t.observers;n.length;)n.shift().complete()}}))},n.prototype.unsubscribe=function(){this.isStopped=this.closed=!0,this.observers=this.currentObservers=null},Object.defineProperty(n.prototype,"observed",{get:function(){var t;return(null===(t=this.observers)||void 0===t?void 0:t.length)>0},enumerable:!1,configurable:!0}),n.prototype._trySubscribe=function(n){return this._throwIfClosed(),t.prototype._trySubscribe.call(this,n)},n.prototype._subscribe=function(t){return this._throwIfClosed(),this._checkFinalizedStatuses(t),this._innerSubscribe(t)},n.prototype._innerSubscribe=function(t){var n=this,r=this,e=r.hasError,o=r.isStopped,i=r.observers;return e||o?T:(this.currentObservers=null,i.push(t),new O((function(){n.currentObservers=null,E(i,t)})))},n.prototype._checkFinalizedStatuses=function(t){var n=this,r=n.hasError,e=n.thrownError,o=n.isStopped;r?t.error(e):o&&t.complete()},n.prototype.asObservable=function(){var t=new st;return t.source=this,t},n.create=function(t,n){return new Xt(t,n)},n}(st),Xt=function(t){function n(n,r){var e=t.call(this)||this;return e.destination=n,e.source=r,e}return y(n,t),n.prototype.next=function(t){var n,r;null===(r=null===(n=this.destination)||void 0===n?void 0:n.next)||void 0===r||r.call(n,t)},n.prototype.error=function(t){var n,r;null===(r=null===(n=this.destination)||void 0===n?void 0:n.error)||void 0===r||r.call(n,t)},n.prototype.complete=function(){var t,n;null===(n=null===(t=this.destination)||void 0===t?void 0:t.complete)||void 0===n||n.call(t)},n.prototype._subscribe=function(t){var n,r;return null!==(r=null===(n=this.source)||void 0===n?void 0:n.subscribe(t))&&void 0!==r?r:T},n}(Wt),$t=function(t){function n(n){var r=t.call(this)||this;return r._value=n,r}return y(n,t),Object.defineProperty(n.prototype,"value",{get:function(){return this.getValue()},enumerable:!1,configurable:!0}),n.prototype._subscribe=function(n){var r=t.prototype._subscribe.call(this,n);return!r.closed&&n.next(this._value),r},n.prototype.getValue=function(){var t=this,n=t.hasError,r=t.thrownError,e=t._value;if(n)throw r;return this._throwIfClosed(),e},n.prototype.next=function(n){t.prototype.next.call(this,this._value=n)},n}(Wt);function tn(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];var r=Rt(t);return At((function(n,e){for(var o=t.length,i=new Array(o),u=t.map((function(){return!1})),c=!1,a=function(n){_t(t[n]).subscribe(It(e,(function(t){i[n]=t,c||u[n]||(u[n]=!0,(c=u.every(ut))&&(u=null))}),H))},s=0;s<o;s++)a(s);n.subscribe(It(e,(function(t){if(c){var n=w([t],x(i));e.next(r?r.apply(void 0,w([],x(n))):n)}})))}))}function nn(t,n){for(var r=[],e=2;e<arguments.length;e++)r[e-2]=arguments[e];if(!0!==n){if(!1!==n){var o=new nt({next:function(){o.unsubscribe(),t()}});return n.apply(void 0,w([],x(r))).subscribe(o)}}else t()}function rn(t,n,r){return void 0===r&&(r=ut),function(e){en(n,(function(){for(var o=t.length,i=new Array(o),u=o,c=o,a=function(o){en(n,(function(){var a=Yt(t[o],n),s=!1;a.subscribe(It(e,(function(t){i[o]=t,s||(s=!0,c--),c||e.next(r(i.slice()))}),(function(){--u||e.complete()})))}),e)},s=0;s<o;s++)a(s)}),e)}}function en(t,n,r){t?Tt(r,t,n):n()}var on=Array.isArray;function un(t){return 1===t.length&&on(t[0])?t[0]:t}function cn(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];var r=Rt(t);return r?ct(cn.apply(void 0,w([],x(t))),Pt(r)):At((function(n,r){rn(w([n],x(un(t))))(r)}))}function an(t,n){return void 0===n&&(n=!1),At((function(r,e){var o=0;r.subscribe(It(e,(function(r){var i=t(r,o++);(i||n)&&e.next(r),!i&&e.complete()})))}))}function sn(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];var r=Ut(t);return Yt(t,r)}var ln,fn,pn,hn,dn,vn={schedule:function(t){var n=requestAnimationFrame,r=cancelAnimationFrame,e=vn.delegate;e&&(n=e.requestAnimationFrame,r=e.cancelAnimationFrame);var o=n((function(n){r=void 0,t(n)}));return new O((function(){return null==r?void 0:r(o)}))},requestAnimationFrame:function(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];var r=vn.delegate;return((null==r?void 0:r.requestAnimationFrame)||requestAnimationFrame).apply(void 0,w([],x(t)))},cancelAnimationFrame:function(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];var r=vn.delegate;return((null==r?void 0:r.cancelAnimationFrame)||cancelAnimationFrame).apply(void 0,w([],x(t)))},delegate:void 0},yn=function(t){function n(n,r){var e=t.call(this,n,r)||this;return e.scheduler=n,e.work=r,e}return y(n,t),n.prototype.requestAsyncId=function(n,r,e){return void 0===e&&(e=0),null!==e&&e>0?t.prototype.requestAsyncId.call(this,n,r,e):(n.actions.push(this),n._scheduled||(n._scheduled=vn.requestAnimationFrame((function(){return n.flush(void 0)}))))},n.prototype.recycleAsyncId=function(n,r,e){var o;if(void 0===e&&(e=0),null!=e?e>0:this.delay>0)return t.prototype.recycleAsyncId.call(this,n,r,e);var i=n.actions;null!=r&&(null===(o=i[i.length-1])||void 0===o?void 0:o.id)!==r&&(vn.cancelAnimationFrame(r),n._scheduled=void 0)},n}(F),bn=function(t){function n(){return null!==t&&t.apply(this,arguments)||this}return y(n,t),n.prototype.flush=function(t){this._active=!0;var n=this._scheduled;this._scheduled=void 0;var r,e=this.actions;t=t||e.shift();do{if(r=t.execute(t.state,t.delay))break}while((t=e[0])&&t.id===n&&e.shift());if(this._active=!1,r){for(;(t=e[0])&&t.id===n&&e.shift();)t.unsubscribe();throw r}},n}(L),mn=new bn(yn),xn={a:{x:-1,y:0},d:{x:1,y:0},w:{x:0,y:-1},s:{x:0,y:1}},wn=xn.d,gn=function(t,n){return t.x===n.x&&t.y===n.y},_n=function(t,n){return Math.floor(Math.random()*(n-t+1)+t)},Sn=function(){return{x:_n(0,19),y:_n(0,19)}},An=pt(300),In=function t(n,r,e,o){if(S(e)&&(o=e,e=void 0),o)return t(n,r,e).pipe(Pt(o));var i=x(function(t){return S(t.addEventListener)&&S(t.removeEventListener)}(n)?Ft.map((function(t){return function(o){return n[t](r,o,e)}})):function(t){return S(t.addListener)&&S(t.removeListener)}(n)?jt.map(Bt(n,r)):function(t){return S(t.on)&&S(t.off)}(n)?zt.map(Bt(n,r)):[],2),u=i[0],c=i[1];if(!u&&ht(n))return kt((function(n){return t(n,r,e)}))(_t(n));if(!u)throw new TypeError("Invalid event target");return new st((function(t){var n=function(){for(var n=[],r=0;r<arguments.length;r++)n[r]=arguments[r];return t.next(1<n.length?n:n[0])};return u(n),function(){return c(n)}}))}(document,"keydown").pipe((ln=An,At((function(t,n){var r=[];return t.subscribe(It(n,(function(t){return r.push(t)}),(function(){n.next(r),n.complete()}))),ln.subscribe(It(n,(function(){var t=r;r=[],n.next(t)}),H)),function(){r=null}}))),Lt((function(t){return t.length>0})),Ot((function(t){return xn[t[t.length-1].key]})),Lt((function(t){return Boolean(t)})),function(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];var r=Ut(t);return At((function(n,e){(r?Dt(t,n,r):Dt(t,n)).subscribe(e)}))}(wn),Vt((function(t,n){return e=n,(r=t).x===-1*e.x||r.y===-1*e.y?t:n;var r,e})),Jt()),En=new $t(0),On=En.pipe(Vt((function(t,n){return t+n}),2)),Tn=An.pipe(tn(In,On,(function(t,n,r){return[n,r]})),Vt((function(t,n){var r=n[0],e=n[1],o=r,i=function(t,n,r){if(r||2===arguments.length)for(var e,o=0,i=n.length;o<i;o++)!e&&o in n||(e||(e=Array.prototype.slice.call(n,0,o)),e[o]=n[o]);return t.concat(e||Array.prototype.slice.call(n))}([{x:t[0].x+o.x,y:t[0].y+o.y}],t,!0);return e<i.length&&i.pop(),i}),function(t){for(var n=[],r=1;r>=0;r--)n.push({x:r,y:0});return n}())),kn=Tn.pipe(Vt((function(t,n){for(var r=t,e=n.some((function(t){return gn(t,r)}));e;)r=Sn(),e=n.some((function(t){return gn(t,r)}));return r}),Sn()),Jt(),function(t){void 0===t&&(t={});var n=t.connector,r=void 0===n?function(){return new Wt}:n,e=t.resetOnError,o=void 0===e||e,i=t.resetOnComplete,u=void 0===i||i,c=t.resetOnRefCountZero,a=void 0===c||c;return function(t){var n,e,i,c=0,s=!1,l=!1,f=function(){null==e||e.unsubscribe(),e=void 0},p=function(){f(),n=i=void 0,s=l=!1},h=function(){var t=n;p(),null==t||t.unsubscribe()};return At((function(t,d){c++,l||s||f();var v=i=null!=i?i:r();d.add((function(){0!=--c||l||s||(e=nn(h,a))})),v.subscribe(d),!n&&c>0&&(n=new nt({next:function(t){return v.next(t)},error:function(t){l=!0,f(),e=nn(p,o,t),v.error(t)},complete:function(){s=!0,f(),e=nn(p,u),v.complete()}}),_t(t).subscribe(n))}))(t)}}());kn.pipe((1,Lt((function(t,n){return 1<=n}))),(fn=function(){En.next(1)},dn=S(fn)?{next:fn,error:pn,complete:hn}:fn,dn?At((function(t,n){var r;null===(r=dn.subscribe)||void 0===r||r.call(dn);var e=!0;t.subscribe(It(n,(function(t){var r;null===(r=dn.next)||void 0===r||r.call(dn,t),n.next(t)}),(function(){var t;e=!1,null===(t=dn.complete)||void 0===t||t.call(dn),n.complete()}),(function(t){var r;e=!1,null===(r=dn.error)||void 0===r||r.call(dn,t),n.error(t)}),(function(){var t,n;e&&(null===(t=dn.unsubscribe)||void 0===t||t.call(dn)),null===(n=dn.finalize)||void 0===n||n.call(dn)})))})):ut)).subscribe();var Cn=Tn.pipe(function(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];return cn.apply(void 0,w([],x(t)))}(kn),an((function(t){var n=t[0],r=n[0];return!(r.x<0||r.y<0||r.x>19||r.y>19||function(t,n){for(var r=!1,e=3;e<n.length;e+=1)gn(t,n[e])&&(r=!0);return r}(r,n))})),function(){for(var t=[],n=0;n<arguments.length;n++)t[n]=arguments[n];return function(n){return Dt(n,sn.apply(void 0,w([],x(t))))}}("end")),Pn=pt(100,mn).pipe(tn(Cn,(function(t,n){return n})),an((function(t){return"end"!==t}))),jn=function(){function t(t){var n=this;this.renderGameOver=function(){n.ctx.font="bold 15px sans-serif",n.ctx.fillStyle="black",n.ctx.textAlign="center",n.ctx.textBaseline="middle",n.ctx.fillText("Game Over",150,90)},this.ctx=t}return t.prototype.getSnakeBoxColor=function(t){return 0===t?"#3f72af":"#c7d0d5"},t.prototype.drawBox=function(t,n){this.ctx.fillStyle=n,this.ctx.fillRect(15*t.x,15*t.y,15,15)},t.prototype.drawBackground=function(){this.ctx.fillStyle="#EEE",this.ctx.fillRect(0,0,300,300)},t.prototype.drawSnake=function(t){var n=this;t.forEach((function(t,r){n.drawBox(t,n.getSnakeBoxColor(r))}))},t.prototype.drawApple=function(t){this.drawBox(t,"#ff7148")},t.prototype.drawScore=function(t){this.ctx.font="bold 80px sans-serif",this.ctx.fillStyle="#d9d9d9",this.ctx.textAlign="center",this.ctx.textBaseline="middle",this.ctx.fillText(t.toString(),150,150)},t.prepareCanvasSize=function(t){t.width=300,t.height=300},t}();const Fn=jn;var zn=document.getElementById("appCan"),Bn=new Fn(zn.getContext("2d"));Fn.prepareCanvasSize(zn),Pn.subscribe({next:function(t){var n=t[0],r=t[1];Bn.drawBackground(),Bn.drawScore(n.length-2),Bn.drawSnake(n),Bn.drawApple(r)},complete:function(){return Bn.renderGameOver()}})})()})();
